refactores BlogSection into its own file and adds it to bootcamp london
