+++ src/pages/graphql/training/bootcamp/index.js
@@ -22,6 +22,7 @@ import { Image } from 'src/components/elements'
 import { Breadcrumb } from 'src/components/navigation'
 import { GRAPHQL_BOOTCAMP } from 'src/config/data'
 import header from 'src/components/layout/Header.json'
+import BlogSection from 'src/components/blog/BlogSection'
 
 const GraphQL = () => (
   <Layout>
@@ -142,8 +143,8 @@ const GraphQL = () => (
               </Row>
             </Grid>
           </Section>
-
           <TrustedBySection />
+          <BlogSection tags={['graphql', 'beginner']} />
           <UpcomingTrainingSection trainings={trainings} />
         </React.Fragment>
       )
