+++ src/pages/react/training/bootcamp/index.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import { StaticQuery, graphql } from 'gatsby'
 
 import { BOOTCAMP } from 'src/../images/imageNames'
 import Layout from 'src/components/layout'
@@ -25,9 +24,7 @@ import { DAVIAN } from 'src/config/images'
 import { Breadcrumb } from 'src/components/navigation'
 import { REACT_BOOTCAMP } from 'src/config/data'
 import header from 'src/components/layout/Header.json'
-import Flex from 'src/components/layout/Flex'
-
-import PostCard from 'src/components/blog/PostCard'
+import BlogSection from 'src/components/blog/BlogSection'
 
 const Bootcamps = props => (
   <Layout>
@@ -148,74 +145,6 @@ const Bootcamps = props => (
       )
     }}
   </Layout>
-)
-
-const BlogSection = ({ tags = [] }) => (
-  <StaticQuery
-    query={graphql`
-      query getPosts($limit: Int = 1000) {
-        allMarkdownRemark(
-          filter: {
-            fields: { slug: { regex: "/(/blog/|/react/|/graphql/)/" } }
-          }
-          sort: { fields: [frontmatter___order], order: DESC }
-          limit: $limit
-        ) {
-          edges {
-            node {
-              fields {
-                slug
-              }
-              frontmatter {
-                title
-                imageUrl
-                tags
-              }
-              excerpt
-            }
-          }
-        }
-      }
-    `}
-    render={data => {
-      const posts = data.allMarkdownRemark.edges
-        .filter(
-          ({
-            node: {
-              frontmatter: { tags: postTags },
-            },
-          }) =>
-            tags &&
-            tags.every(
-              constraint => postTags && postTags.some(tag => tag === constraint)
-            )
-        )
-        .slice(0, 3)
-
-      if (!posts || !posts.length) {
-        return null
-      }
-
-      return (
-        <Section>
-          <Grid>
-            <Row>
-              <Col>
-                <H2>Related posts</H2>
-              </Col>
-            </Row>
-            <Row>
-              {posts.map(({ node: post }) => (
-                <Col md={4} key={post.fields.slug}>
-                  <PostCard post={post} />
-                </Col>
-              ))}
-            </Row>
-          </Grid>
-        </Section>
-      )
-    }}
-  />
 )
 
 export default Bootcamps
